# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AdaptLearn API
  version: 0.1.0
  summary: Backend contract for AdaptLearn v1 (Phases 2-5).
  description: |
    Source of truth for AdaptLearn HTTP contracts. Pydantic models under
    `backend/contracts/` are generated from this document via
    `backend/scripts/gen_contracts.py`. Edit YAML, then regen.

    See `docs/superpowers/specs/2026-05-03-adaptlearn-v1-design.md`
    sections 3.3, 3.4, 4.1-4.4, 5 for upstream design constraints.
servers:
  - url: http://localhost:8000
    description: Local development backend
...
paths:
...
  /api/profile/aggregate:
    get:
      tags: [profile]
      summary: Cross-session aggregate profile dashboard.
      description: |
        Aggregates TopicProfile fields and LearningEvent counts across every
        session for the given user. Pure SQL/Python — no LLM calls.
      operationId: getAggregateProfile
      parameters:
        - in: query
          name: user_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Aggregate snapshot across all sessions for the user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregateProfileResponse"
  /api/profile/{session_id}:
    get:
      tags: [profile]
      summary: Read a session's topic profile plus recent learning events.
      operationId: getProfile
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Profile snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    SessionId:
      in: path
      name: session_id
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    KnowledgeLevel:
      type: string
      enum: [beginner, intermediate, advanced]
...
    TopicProfile:
      type: object
      additionalProperties: false
      properties:
        knowledge_level:
          oneOf:
            - $ref: "#/components/schemas/KnowledgeLevel"
            - type: "null"
          default: null
        confirmed_gaps:
          type: array
          items: { type: string }
          default: []
        mastered_concepts:
          type: array
          items: { type: string }
          default: []
        focus_target_gap:
          type: [string, "null"]
          default: null
        last_session_summary:
          type: [string, "null"]
          default: null
...
    LearningEventResponse:
      type: object
      additionalProperties: false
      required: [id, session_id, gap_tested, question, correct, created_at]
      properties:
        id:          { type: integer }
        session_id:  { type: string }
        gap_tested:  { type: string }
        question:    { type: string }
        correct:     { type: boolean }
        created_at:  { type: string, format: date-time }

    # ---- Tool argument schemas (LiteLLM tool params) ----
...
    ProfileResponse:
      type: object
      additionalProperties: false
      required: [profile, recent_learning_events]
      properties:
        profile: { $ref: "#/components/schemas/TopicProfile" }
        recent_learning_events:
          type: array
          items: { $ref: "#/components/schemas/LearningEventResponse" }
    AggregateConceptCount:
      type: object
      additionalProperties: false
      required: [concept, count, first_seen_session_id]
      description: |
        Concept (mastered or gap) plus how many sessions it appears in and the
        earliest session id where it was recorded.
      properties:
        concept: { type: string }
        count: { type: integer, minimum: 1 }
        first_seen_session_id: { type: string }
    KnowledgeLevelDistribution:
      type: object
      additionalProperties: false
      required: [beginner, intermediate, advanced, unknown]
      description: |
        Count of sessions per knowledge_level (unknown = null knowledge_level).
      properties:
        beginner: { type: integer, minimum: 0 }
        intermediate: { type: integer, minimum: 0 }
        advanced: { type: integer, minimum: 0 }
        unknown: { type: integer, minimum: 0 }
    RecentSessionSummary:
      type: object
      additionalProperties: false
      required: [id, topic, created_at]
      description: Compact session header for the "recent topics" widget.
      properties:
        id: { type: string }
        topic: { type: string }
        created_at: { type: string, format: date-time }
        ended_at: { type: [string, "null"], format: date-time, default: null }
    AggregateProfileResponse:
      type: object
      additionalProperties: false
      required:
        - total_sessions
        - active_sessions
        - ended_sessions
        - total_learning_events
        - combined_mastered_concepts
        - combined_confirmed_gaps
        - knowledge_level_distribution
        - recent_topics
      description: |
        Cross-session aggregate view powering the top-level /profile dashboard.
        All arrays are sorted by count desc (mastered + gaps) or created_at desc
        (recent_topics).
      properties:
        total_sessions: { type: integer, minimum: 0 }
        active_sessions: { type: integer, minimum: 0 }
        ended_sessions: { type: integer, minimum: 0 }
        total_learning_events: { type: integer, minimum: 0 }
        last_active_at:
          type: [string, "null"]
          format: date-time
          default: null
        combined_mastered_concepts:
          type: array
          items: { $ref: "#/components/schemas/AggregateConceptCount" }
        combined_confirmed_gaps:
          type: array
          items: { $ref: "#/components/schemas/AggregateConceptCount" }
        knowledge_level_distribution:
          $ref: "#/components/schemas/KnowledgeLevelDistribution"
        recent_topics:
          type: array
          items: { $ref: "#/components/schemas/RecentSessionSummary" }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail: { type: string }